When out of stock, Email should be required when customer not logged in
Make the consistent check whether the email addres should be possible to provide by user - the actions.php controller checks if customer is logged in to decide wheter to take email from parameter or from customer context. This change prevents the scenario whit missing email information when customer has the account but is not logged in when making the request.